LPS-117566 Move to provider package
