+++ modules/apps/content-dashboard/content-dashboard-api/bnd.bnd
@@ -3,4 +3,5 @@ Bundle-SymbolicName: com.liferay.content.dashboard.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.content.dashboard.item.action,\
-	com.liferay.content.dashboard.item.action.exception
+	com.liferay.content.dashboard.item.action.exception,\
+	com.liferay.content.dashboard.item.action.provider
