+++ modules/apps/content-dashboard/content-dashboard-api/src/main/java/com/liferay/content/dashboard/item/action/provider/ContentDashboardItemActionProvider.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.content.dashboard.item.action;
+package com.liferay.content.dashboard.item.action.provider;
 
+import com.liferay.content.dashboard.item.action.ContentDashboardItemAction;
 import com.liferay.content.dashboard.item.action.exception.ContentDashboardItemActionException;
 
 import javax.servlet.http.HttpServletRequest;
