+++ modules/apps/content-dashboard/content-dashboard-api/src/main/resources/com/liferay/content/dashboard/item/action/provider/packageinfo
