+++ modules/dxp/apps/analytics/analytics-reports-api/bnd.bnd
@@ -2,5 +2,5 @@ Bundle-Name: Liferay Analytics Reports API
 Bundle-SymbolicName: com.liferay.analytics.reports.api
 Bundle-Version: 1.1.0
 Export-Package:\
-	com.liferay.analytics.reports.info.action,\
+	com.liferay.analytics.reports.info.action.provider,\
 	com.liferay.analytics.reports.info.item
