+++ modules/dxp/apps/analytics/analytics-reports-api/src/main/resources/com/liferay/analytics/reports/info/action/provider/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
